% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.evsi.N.R
\name{comp.evsi.N}
\alias{comp.evsi.N}
\title{Function generates the data points that will be used to calculate the EVSI for different
sample sizes.}
\usage{
comp.evsi.N(model.stats, data, N, N.range = c(30, 1500), effects, costs,
  he = NULL, parameters = NULL, evi = NULL, Q = 50, length.wtp = 150,
  data.stats = NULL, update = c("bugs", "jags"), n.burnin = 1000,
  n.thin = 1, n.iter = 5000)
}
\arguments{
\item{model.stats}{A .txt file containing the model file of a Bayesian model.
This should be a BUGS or JAGS model.}

\item{data}{A string or vector of strings that defines the name of the future
data in the model file. If the future data has already been generated then
the data arguement can be given as a list. The elements of the data list should
be data lists for JAGS or BUGS.}

\item{N}{A string defining the name of the variable controlling the sample size in the
model.}

\item{N.range}{A two-vector defining the minimum and maximum values considered for the
sample size. If the future data is given in the data arguement then N.range should
give the sample sizes for which the data was generated.}

\item{effects}{This can either be given as a string which defines the name of the
effectivness measure in the BUGS/JAGS model. Or it can be given as a
function that takes the output of the model and calculates the
effectiveness measure. The inputs of this function should be parameter
names found in the model file.}

\item{costs}{This has a similar format to e but gives the value of the costs in
the model file or a function defining the costs.}

\item{he}{A bcea object containing the base case analysis for the model.}

\item{parameters}{A list of the names of the parameters that are being targetted by the study.
This can be given OR an evppi object needs to be given.}

\item{evi}{An evppi object that contains the EVPPI analysis for the parameters
that are being informed by the sample information.}

\item{Q}{The number of quadrature points used the estimate the EVSI.}

\item{length.wtp}{The number of willingness to pay values that should be considered to estimate the EVSI}

\item{data.stats}{A data file for the BUGS/JAGS model. This is the data used to inform
the base case analysis. If empty then it is assumed that the models are}

\item{update}{Defines the Bayesian engine that should be used to update the
the Bayesian model file given in model.stats.}

\item{n.burnin}{The burnin for the JAGS/BUGS model}

\item{n.thin}{The thinning for the JAGS/BUGS model}

\item{n.iter}{The number of interations for the JAGS/BUGS model

OUTPUTS.}
}
\value{
An evsi.comp object to be used in the evsi.calc function.
}
\description{
Function generates the data points that will be used to calculate the EVSI for different
sample sizes.
}
