% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.calc.R
\name{evsi.calc}
\alias{evsi.calc}
\title{Calculates the EVSI accross willingness-to-pay, sample size and gives uncertainty for these
measurements as required. This function can be fed into/used in the other plot functions
This should allow people to upload their own EVSI matrices and still use the plotting capabilities.}
\usage{
evsi.calc(mm.var, wtp = NULL, N = NULL, CI = NULL)
}
\arguments{
\item{mm.var}{An mm.var object computed from the mm.post.var function in the EVSI package.
The posterior variances can also be calculated external to the package and passed as a list
to the evsi.calc function. If this is the case a bcea object, an evppi object, N.calc and
a Bayesian updating program, if mm.var is estimated across different sample sizes must 
also be provided.}

\item{wtp}{The willingness to pay value(s) for which the EVSI should be calculated.
It will be chosen as all the wtp values in the bcea object, either passed as he or 
directly from the mm.var object.}

\item{N}{The sample sizes for which the EVSI should be calculated. If NULL it will select
a range of values for which the EVSI was calculated in the mm.var object. If the mm.var
object is not given then this argument must be provided.}

\item{CI}{If EVSI is calculated across multiple sample sizes then this gives the confidence 
level for the credible intervals of the EVSI. If NULL chosen as 95% and 75% intervals.}
}
}
\value{
An evsi object.
1. evsi An array containing the EVSI by wtp, N and across different uncertaincies
2. attrib A list of wtp, N and prob describing the attributes of the evsi matrix.
3. evppi An evppi object containing all the information about the calculation of
  EVPPI.
4. he A bcea object containing all the information about the underlying health
  economic model
}
\description{
Calculates the EVSI accross willingness-to-pay, sample size and gives uncertainty for these
measurements as required. This function can be fed into/used in the other plot functions
This should allow people to upload their own EVSI matrices and still use the plotting capabilities.
}
