% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.plot.dynam.R
\name{evsi.plot.dynam}
\alias{evsi.plot.dynam}
\title{Function launches shiny app to show dynamic graphics
INPUTS}
\usage{
evsi.plot.dynam(evsi, width = 600, height = 600)
}
\arguments{
\item{evsi}{An evsi object for which dynamic graphics are needed.}

\item{width}{The width of the graph produced in the webapp (default 600pt)}

\item{height}{The height of the graph produced in the webapp (default 600pt)}
}
\value{
A dynamic shiny app that plots the EVSI for different wtp,
sample size, incidence population, time horizon and costs as required.
}
\description{
Function launches shiny app to show dynamic graphics
INPUTS
}
