% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.pop.R
\name{evsi.pop}
\alias{evsi.pop}
\title{Produces a graphic that considers the cost-effectiveness of a single future trial for different
Time Horizons for the treatment and incidence population
INPUTS:}
\usage{
evsi.pop(evsi, trial.cost = NULL, setup = NULL, pp = NULL, Pop = c(0,
  10000), Time = c(1, 20), Dis = 0.035, wtp = NULL, N = NULL,
  pos = c("topright"))
}
\arguments{
\item{evsi}{An evsi object.}

\item{trial.cost}{Either a fixed trial cost (numeric) or a distribution for a trial cost (character string)
- to give probability of cost-effectiveness for the trial. Can be NULL if setup and pp are both given}

\item{setup}{Gives the setup costs of the trial. Can be NULL if trial.costs are given. If NULL and pp given
setup is taken to be 0.}

\item{pp}{Gives the per person costs of the trial. Can be NULL if trial.costs are given.}

\item{Pop}{The minimum and maximum possible values for the number of people benefitting from
the treatment}

\item{Time}{The minimum and maximum possible values for the time horizon of the treatment.}

\item{Dis}{The discount rate for future studies, based on NICE guidelines.}

\item{wtp}{The willingness to pay value that this analysis is being undetaken for.
If NULL then the function will automatically select the central wtp to default in BCEA will be 25000}

\item{N}{The sample size which we want to consider for the trial. If NULL then taken as the median trial size
considered.}

\item{pos}{The position where the legend will be printed in the resulting graph.}
}
\value{
A graphic that gives the probability of a cost-effective trial.
}
\description{
Produces a graphic that considers the cost-effectiveness of a single future trial for different
Time Horizons for the treatment and incidence population
INPUTS:
}
