% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.R
\name{evsi}
\alias{evsi}
\title{Compute the EVSI for a fixed design using the Heath et al. method}
\usage{
evsi(model.stats, data, effects = NULL, costs = NULL, he = NULL,
  evi = NULL, parameters = NULL, Q = 30, data.stats = NULL,
  update = c("bugs", "jags"), n.burnin = 1000, n.thin = 1,
  n.iter = 5000)
}
\arguments{
\item{model.stats}{A .txt file containing the model file of a Bayesian model.
This should be a BUGS or JAGS model.}

\item{data}{A string or vector of strings that defines the name of the future
data in the model file. If the future data has already been generated then
the data arguement can be given as a list. The elements of the data list should
be data lists for JAGS or BUGS.}

\item{effects}{This can either be given as a string which defines the name of the
effectivness measure in the BUGS/JAGS model. Or it can be given as a
function that takes the output of the model and calculates the
effectiveness measure. The inputs of this function should be parameter
names found in the model file.}

\item{costs}{This has a similar format to e but gives the value of the costs in
the model file or a function defining the costs.}

\item{he}{A bcea object containing the base case analysis for the model.}

\item{evi}{An evppi object that contains the EVPPI analysis for the parameters
that are being informed by the sample information.}

\item{parameters}{A list of the names of the parameters that are being targetted by the study.
This can be given OR an evppi object needs to be given.}

\item{Q}{The number of quadrature points used the estimate the EVSI.}

\item{data.stats}{A data file for the BUGS/JAGS model. This is the data used to inform
the base case analysis. If empty then it is assumed that the models are}

\item{update}{Defines the Bayesian engine that should be used to update the
the Bayesian model file given in model.stats.}

\item{n.burnin}{The burnin for the JAGS/BUGS model}

\item{n.thin}{The thinning for the JAGS/BUGS model}

\item{n.iter}{The number of interations for the JAGS/BUGS model}
}
\value{
An evsi object.
1. evsi An array containing the EVSI by wtp, N and across different uncertaincies
2. attrib A list of wtp, N and prob describing the attributes of the evsi matrix.
3. evppi An evppi object containing all the information about the calculation of
  EVPPI.
4. he A bcea object containing all the information about the underlying health
  economic model
}
\description{
Compute the EVSI for a fixed design using the Heath et al. method
}
\examples{
NULL
}
