% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.upload.R
\name{evsi.upload}
\alias{evsi.upload}
\title{Uploads EVSI either from a matrix in R or a .csv file.
Allows EVSI calculated outside of package to be uploaded so graphics can be used.
INPUTS}
\usage{
evsi.upload(e, c, parameter, input, EVSI.mat = NULL, model.stats = NULL,
  wtp = NULL, N = NULL)
}
\arguments{
\item{e}{A matrix containing simuations for the variable of clinical effectiveness for
each inteventions being considered.}

\item{c}{A matrix containing simuations for the variable of cost for each inteventions
being considered.}

\item{parameter}{A vector of parameters that are updated by the future dataset. This can
be given as a string of names or a numeric vector.}

\item{input}{A matrix of PSA simulations from the underlying health economic model.}

\item{EVSI.mat}{A matrix of EVSI values. The COLUMNS relate the different willingness
to pay values and the ROWS relate the different sample sizes.}

\item{model.stats}{A .csv file containing the matrix of EVSI values. Again the COLUMNS
should relate the different WTP values and the ROWS to different sample sizes.}

\item{wtp}{A vector of willingness to pay values for which the EVSI was calculated.
If NULL then the first row of the EVSI matrix is assumed to contain the WTP values.}

\item{N}{A vector of sample sizes for which the EVSI is calculated. If NULL then the
first column of the EVSI matrix is assumed to contain the sample sizes.}
}
\value{
An evsi object.
1. evsi An array containing the EVSI by wtp, N and across different uncertaincies
2. attrib A list of wtp, N and prob describing the attributes of the evsi matrix.
3. evppi An evppi object containing all the information about the calculation of
  EVPPI.
4. he A bcea object containing all the information about the underlying health
  economic model
  @examples
  ...
}
\description{
Uploads EVSI either from a matrix in R or a .csv file.
Allows EVSI calculated outside of package to be uploaded so graphics can be used.
INPUTS
}
