% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.quantiles.R
\name{gen.quantiles}
\alias{gen.quantiles}
\title{Compute the posterior variance across Q difference proposed data collection exercises 
This can be used to estimate the EVSI using the Heath et al. method.}
\usage{
gen.quantiles(parameter, param.mat, Q, N.size = NULL)
}
\arguments{
\item{parameter}{A vector of parameters that define the sampling distribution of the future
data. This can be given as a string (or vector of strings) of names or a numeric vector, 
corresponding to the column numbers of parameters of interest.}

\item{param.mat}{A matrix containing the simulations for the parameters. The matrix should 
have the column names matching the names of the parameters.}

\item{Q}{The number of quadrature points used the estimate the EVSI.}

\item{N.size}{If the EVSI is to be computed across sample size, N.size is a vector that 
defines the maximum and minimum sample sizes for the proposed future study.
If NULL then the the sample size of the future data will not be proposed by the gen.quantiles
function.}
}
\value{
A dataframe containing the parameters simulations and sample size that can
 be used to generate future data to compute the EVSI using the Heath et al. moment
 matching EVSI computation method.
}
\description{
Compute the posterior variance across Q difference proposed data collection exercises 
This can be used to estimate the EVSI using the Heath et al. method.
}
\examples{
NULL
}
