% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi.plot.dynam.R
\name{launch.App}
\alias{launch.App}
\title{Function launches shiny app to show dynamic graphics
INPUTS}
\usage{
launch.App(...)
}
\arguments{
\item{...}{Either empty or an evsi object. If given an evsi object then graphics will be displayed
If empty then you can upload the EVSI using an upload page.}
}
\value{
A dynamic shiny app that plots the EVSI for different wtp,
sample size, incidence population, time horizon and costs as required.
}
\description{
Function launches shiny app to show dynamic graphics
INPUTS
}
