% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.enbs.R
\name{plot.enbs}
\alias{plot.enbs}
\title{Produces a graphic that considers the cost-effectiveness of a single future trial for different
Time Horizons for the treatment and incidence population
INPUTS:}
\usage{
\method{plot}{enbs}(evsi, setup, pp, prob = NULL, Pop = 10000,
  Time = 10, Dis = 0.035, wtp = NULL, N = NULL,
  pos = c("bottomright"))
}
\arguments{
\item{evsi}{An evsi object.}

\item{setup}{Gives the minimum and maximum possible setup costs of the trial.}

\item{pp}{Gives the minimum and maxiumum possible per person costs of the trial.}

\item{prob}{The credible interval bounds that you would like plotted.
By default the CI intervals.}

\item{Pop}{The number of people benefitting from the treatment.}

\item{Time}{The minimum and maximum possible values for the time horizon of the treatment.}

\item{Dis}{The discount rate for future studies, based on NICE guidelines.}

\item{wtp}{The willingness to pay value that this analysis is being undetaken for.
If NULL then the function will automatically select the central wtp to default in BCEA will be 25000}

\item{N}{The sample sizes for which we want to consider the ENBS. If NULL taken as all the
sample sizes that the EVSI has been calculated for.}

\item{pos}{The position where the legend will be printed in the resulting graph.}
}
\value{
A graphic that gives the ENBS for the range of N values given.
}
\description{
Produces a graphic that considers the cost-effectiveness of a single future trial for different
Time Horizons for the treatment and incidence population
INPUTS:
}
