% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.evsi.N.R
\name{plot.evsi.N}
\alias{plot.evsi.N}
\title{Calculating the EVSI for a specific WTP giving the uncertainty bands across the different
samples sizes}
\usage{
\method{plot}{evsi.N}(evsi, wtp = NULL, pos = c("bottomright"), CI = NULL)
}
\arguments{
\item{evsi}{Output of the comp.evsi.N function}

\item{wtp}{The willingness to pay value that the graphic should be produced for - it will
be chosen if wtp=NULL.}

\item{pos}{The position where the legend will be printed (default='bottomright')}

\item{CI}{The indexes that we would like to take from the CI in the evsi object.}
}
\value{
EVSI The EVSI calculated for a specific wtp with uncertainty estimates.

A graphical representation of the uncertainty.
}
\description{
Calculating the EVSI for a specific WTP giving the uncertainty bands across the different
samples sizes
}
