% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evsi functions.R
\name{plot.evsi}
\alias{plot.evsi}
\title{Plots the EVSI across different willingness-to-pay thresholds for a single design.}
\usage{
\method{plot}{evsi}(evsi, pos = c(0, 0.8), N = NULL)
}
\arguments{
\item{evsi}{An evsi object that calculates the EVSI for a single design using the evsi function}

\item{pos}{An indcation where the legend should be printed}

\item{N}{A vector of values for which the EVSI should be plotted.}
}
\value{
Produces a graphic containing the EVPI, the EVPPI for the key parameters and the EVSI for the
fixed design.
}
\description{
Plots the EVSI across different willingness-to-pay thresholds for a single design.
}
