% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.evsi.R
\name{print.evsi}
\alias{print.evsi}
\title{Summarises the output of an evsi object}
\usage{
\method{print}{evsi}(obj.evsi)
}
\arguments{
\item{obj.evsi}{An evsi object.}
}
\value{
Prints the Value of Information analysis
}
\description{
Summarises the output of an evsi object
}
\author{
Anna Heath and Gianluca Baio
}
