% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.evsi.R
\name{summary.evsi}
\alias{summary.evsi}
\title{Summarises the output of an evsi object}
\usage{
\method{summary}{evsi}(obj.evsi, wtp = NULL, N = NULL, CI = NULL)
}
\arguments{
\item{obj.evsi}{An evsi object.}

\item{wtp}{A fixed willingness-to-pay for which we would like to summarise the
EVSI. If NULL then the willingness-to-pay assocaited with the optimal decision
uncertainty is chosen.}

\item{N}{The sample size for which the EVSI is to be summerised. If NULL then chose
the middle value for N.}
}
\value{
Prints a summary table with information on the Value of Information analysis
}
\description{
Summarises the output of an evsi object
}
\author{
Anna Heath and Gianluca Baio
}
